Address unitary scaling of positions.
We utilize unitary scaling almost everywhere when setting positions;
this is even built into the `YTPositionTrait` object.  However, we don't
utilize this in the setting of the left and right edges of the objects,
because our `PartitionedGrid` objects don't have units associated with
their attributes.

One side effect of this is that when loading a dataset where the unitary
scale is wildly different from the `code_length` scale, our camera
position is strange and set in the wrong absolute coordinate values.